feat: add performance metrics to RNPerfMonitorView



# message auto-generated for no-merge-commit merge:
merge feat/add-perf-metrics into 0.77.1-rc.1-ohos

feat: add performance metrics to RNPerfMonitorView

Created-by: pafry
Commit-by: Patryk Fryda
Merged-by: huangyouhua
Description: [![Screenshot 2025-09-12 at 15.03.09.png](https://raw.gitcode.com/user-images/assets/4370111/1367a4a8-3eb9-46b7-a06b-5c283c478dd1/Screenshot_2025-09-12_at_15.03.09.png 'Screenshot 2025-09-12 at 15.03.09.png')](url)

- Includes an example of using performance metrics based on the existing RNOHMarker (to be replaced with the getter)
- Includes new metrics: views, JS VM, and PSS


This PR is based on the following PRs:
- https://gitcode.com/openharmony-sig/ohos_react_native/pull/1522,
- https://gitcode.com/openharmony-sig/ohos_react_native/pull/1556
These should be merged first.

## Why are these changes being made?
This changes add performance metrics to Perf Monitor

## Changelog
- added performance metrics to RNPerfMonitorView

## Test Plan
https://docs.google.com/document/d/1MJOkU1CJp664stimY0-YsRvSky2vcJ1nSsPpvy8agqw/edit?tab=t.0


## Checklist
- [x] `pnpm verify` command has passed successfully.
- [x] No failures occurred in automatic tests on the `CONCURRENT TESTER` page in the tester app.
- [x] No failures occurred in automatic tests on the `SEQUENTIAL TESTER` page in the tester app.
- [x] MR does not include a breaking change or a review has been requested from a Huawei reviewer.


See merge request: openharmony-sig/ohos_react_native!1558

+++ packages/tester/harmony/react_native_openharmony/src/main/cpp/RNOH/RNInstance.h
@@ -116,7 +116,7 @@ class RNInstance {
    * @brief Returns the unique ID of the instance.
    * @return Instance identifier.
    */
-  virtual int getId() = 0;
+  virtual size_t getId() = 0;
 
   /**
    * @brief Returns the context container which holds shared runtime
